Fix plural agreement and gender agreement.
The word "thread" is referred as a feminine word in portuguese, so the words "um","o" and "liberado" is changed accordingly.
In the phrase "As operações assíncronas também pode aumentar a taxa de transferência em um aplicativo Web,...", the word "pode" refers to the plural expression "as operações assíncronas". So it is changed to "podem".